# contracts/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "CollectionEnvelope_UsageSummaryResource_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/UsageSummaryResource"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "CollectionEnvelope[UsageSummaryResource]",
        "type": "object"
      },
      "CostSummaryResource": {
        "additionalProperties": false,
        "properties": {
          "active_reservations": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Active Reservations",
            "type": "string"
          },
          "actual_cost": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Actual Cost",
            "type": "string"
          },
          "adjustments": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Adjustments",
            "type": "string"
          },
          "currency": {
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "from_time": {
            "format": "date-time",
            "title": "From Time",
            "type": "string"
          },
          "net_provider_cost": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Net Provider Cost",
            "type": "string"
          },
          "organization_id": {
            "format": "uuid",
            "title": "Organization Id",
            "type": "string"
          },
          "project_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "reversals": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Reversals",
            "type": "string"
          },
          "to_time": {
            "format": "date-time",
            "title": "To Time",
            "type": "string"
          },
          "unknown_cost_entries": {
            "minimum": 0.0,
            "title": "Unknown Cost Entries",
            "type": "integer"
          }
        },
        "required": [
          "organization_id",
          "currency",
          "actual_cost",
          "adjustments",
          "reversals",
          "net_provider_cost",
          "active_reservations",
          "unknown_cost_entries",
          "from_time",
          "to_time"
        ],
        "title": "CostSummaryResource",
        "type": "object"
      },
...
      "DataEnvelope_CostSummaryResource_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CostSummaryResource"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "DataEnvelope[CostSummaryResource]",
        "type": "object"
      },
...
      "PageMeta": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "PageMeta",
        "type": "object"
      },
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ProblemField"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code",
          "request_id"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "ProblemField": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "ProblemField",
        "type": "object"
      },
...
      "ResponseMeta": {
        "additionalProperties": false,
        "properties": {
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta",
        "type": "object"
      },
...
      "UsageSummaryResource": {
        "additionalProperties": false,
        "properties": {
          "from_time": {
            "format": "date-time",
            "title": "From Time",
            "type": "string"
          },
          "metric": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Metric",
            "type": "string"
          },
          "organization_id": {
            "format": "uuid",
            "title": "Organization Id",
            "type": "string"
          },
          "project_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "quantity": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity",
            "type": "string"
          },
          "to_time": {
            "format": "date-time",
            "title": "To Time",
            "type": "string"
          },
          "unit": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "organization_id",
          "metric",
          "quantity",
          "unit",
          "from_time",
          "to_time"
        ],
        "title": "UsageSummaryResource",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Versioned product API contract for LUMI AI Design OS.",
    "title": "LUMI AI Design OS API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/costs/summary": {
      "get": {
        "operationId": "getCostSummary",
        "parameters": [
          {
            "in": "query",
            "name": "from_time",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "From Time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to_time",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "To Time",
              "type": "string"
            }
          },
          {
            "description": "Tenant selector. This header is never authorization by itself; NODE-16 must validate the authenticated actor has membership in this organization.",
            "in": "header",
            "name": "X-Lumi-Organization-Id",
            "required": true,
            "schema": {
              "description": "Tenant selector. This header is never authorization by itself; NODE-16 must validate the authenticated actor has membership in this organization.",
              "format": "uuid",
              "title": "X-Lumi-Organization-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_CostSummaryResource_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Precondition Required"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Cost Summary",
        "tags": [
          "costs"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/costs": {
      "get": {
        "operationId": "getProjectCostSummary",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from_time",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "From Time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to_time",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "To Time",
              "type": "string"
            }
          },
          {
            "description": "Tenant selector. This header is never authorization by itself; NODE-16 must validate the authenticated actor has membership in this organization.",
            "in": "header",
            "name": "X-Lumi-Organization-Id",
            "required": true,
            "schema": {
              "description": "Tenant selector. This header is never authorization by itself; NODE-16 must validate the authenticated actor has membership in this organization.",
              "format": "uuid",
              "title": "X-Lumi-Organization-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_CostSummaryResource_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Precondition Required"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Project Cost Summary",
        "tags": [
          "costs"
        ]
      }
    },
...
    "/api/v1/usage": {
      "get": {
        "operationId": "getUsageSummary",
        "parameters": [
          {
            "in": "query",
            "name": "from_time",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "From Time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to_time",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "To Time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "description": "Tenant selector. This header is never authorization by itself; NODE-16 must validate the authenticated actor has membership in this organization.",
            "in": "header",
            "name": "X-Lumi-Organization-Id",
            "required": true,
            "schema": {
              "description": "Tenant selector. This header is never authorization by itself; NODE-16 must validate the authenticated actor has membership in this organization.",
              "format": "uuid",
              "title": "X-Lumi-Organization-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionEnvelope_UsageSummaryResource_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Precondition Required"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Usage Summary",
        "tags": [
          "costs"
        ]
      }
    }
  }
}